+++ pybop/samplers/base_sampler.py
@@ -4,9 +4,9 @@ import numpy as np
 import pints
 import scipy
 
+from pybop import plot
 from pybop._logging import Logger
 from pybop._result import Result
-from pybop.plot import PlotlyManager
 from pybop.problems.problem import Problem
 
 
@@ -168,7 +168,6 @@ class SamplingResult(Result):
         self.chains = chains
         self.all_samples = np.concatenate(chains, axis=0)
         self.num_parameters = self.chains.shape[2]
-        self.go = PlotlyManager().go
 
     def signif(self, x, p: int):
         """
@@ -225,113 +224,25 @@ class SamplingResult(Result):
         """
         Plot trace plots for the posterior samples.
         """
-
-        for i in range(self.num_parameters):
-            fig = self.go.Figure()
-
-            for j, chain in enumerate(self.chains):
-                fig.add_trace(
-                    self.go.Scatter(y=chain[:, i], mode="lines", name=f"Chain {j}")
-                )
-
-            fig.update_layout(
-                title=f"Parameter {i} Trace Plot",
-                xaxis_title="Sample Index",
-                yaxis_title="Value",
-            )
-            fig.update_layout(**kwargs)
-            fig.show()
+        return plot.trace(result=self, **kwargs)
 
     def plot_chains(self, **kwargs):
         """
         Plot posterior distributions for each chain.
         """
-        fig = self.go.Figure()
-
-        for i, chain in enumerate(self.chains):
-            for j in range(chain.shape[1]):
-                fig.add_trace(
-                    self.go.Histogram(
-                        x=chain[:, j],
-                        name=f"Chain {i} - Parameter {j}",
-                        opacity=0.75,
-                    )
-                )
-
-                fig.add_shape(
-                    type="line",
-                    x0=self.mean[j],
-                    y0=0,
-                    x1=self.mean[j],
-                    y1=self.max[j],
-                    name=f"Mean - Parameter {j}",
-                    line=dict(color="Black", width=1.5, dash="dash"),
-                )
-
-        fig.update_layout(
-            barmode="overlay",
-            title="Posterior Distribution",
-            xaxis_title="Value",
-            yaxis_title="Density",
-        )
-        fig.update_layout(**kwargs)
-        fig.show()
+        return plot.chains(result=self, **kwargs)
 
     def plot_posterior(self, **kwargs):
         """
         Plot the summed posterior distribution across chains.
         """
-        fig = self.go.Figure()
+        return plot.posterior(result=self, **kwargs)
 
-        for j in range(self.all_samples.shape[1]):
-            histogram = self.go.Histogram(
-                x=self.all_samples[:, j],
-                name=f"Parameter {j}",
-                opacity=0.75,
-            )
-            fig.add_trace(histogram)
-            fig.add_vline(
-                x=self.mean[j], line_width=3, line_dash="dash", line_color="black"
-            )
-
-        fig.update_layout(
-            barmode="overlay",
-            title="Posterior Distribution",
-            xaxis_title="Value",
-            yaxis_title="Density",
-        )
-        fig.update_layout(**kwargs)
-        fig.show()
-        return fig
-
-    def summary_table(self):
+    def summary_table(self, **kwargs):
         """
         Display summary statistics in a table.
         """
-        summary_stats = self.get_summary_statistics()
-
-        header = ["Statistic", "Value"]
-        values = [
-            ["Mean", summary_stats["mean"]],
-            ["Median", summary_stats["median"]],
-            ["Standard Deviation", summary_stats["std"]],
-            ["95% CI Lower", summary_stats["ci_lower"]],
-            ["95% CI Upper", summary_stats["ci_upper"]],
-        ]
-
-        fig = self.go.Figure(
-            data=[
-                self.go.Table(
-                    header=dict(values=header),
-                    cells=dict(
-                        values=[[row[0] for row in values], [row[1] for row in values]]
-                    ),
-                )
-            ]
-        )
-
-        fig.update_layout(title="Summary Statistics")
-        fig.show()
+        return plot.summary_table(result=self, **kwargs)
 
     def autocorrelation(self, x: np.ndarray) -> np.ndarray:
         """